+++ packages/Angular/Explorer/entity-form-dialog/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log - @memberjunction/ng-entity-form-dialog
 
+## 2.70.0
+
+### Patch Changes
+
+- Updated dependencies [6f74409]
+- Updated dependencies [c9d86cd]
+  - @memberjunction/global@2.70.0
+  - @memberjunction/ng-base-forms@2.70.0
+  - @memberjunction/ng-container-directives@2.70.0
+  - @memberjunction/core@2.70.0
+  - @memberjunction/core-entities@2.70.0
+  - @memberjunction/ng-shared@2.70.0
+
 ## 2.69.1
 
 ### Patch Changes