+++ packages/Angular/Explorer/user-view-properties/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log - @memberjunction/ng-user-view-properties
 
+## 2.70.0
+
+### Patch Changes
+
+- Updated dependencies [6f74409]
+- Updated dependencies [c9d86cd]
+  - @memberjunction/global@2.70.0
+  - @memberjunction/ng-base-forms@2.70.0
+  - @memberjunction/ng-find-record@2.70.0
+  - @memberjunction/ng-resource-permissions@2.70.0
+  - @memberjunction/core@2.70.0
+  - @memberjunction/core-entities@2.70.0
+  - @memberjunction/ng-shared@2.70.0
+
 ## 2.69.1
 
 ### Patch Changes