+++ packages/Angular/Generic/action-gallery/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @memberjunction/ng-action-gallery
 
+## 2.70.0
+
+### Patch Changes
+
+- @memberjunction/ng-container-directives@2.70.0
+- @memberjunction/core@2.70.0
+- @memberjunction/core-entities@2.70.0
+- @memberjunction/ng-ai-test-harness@2.70.0
+
 ## 2.69.1
 
 ### Patch Changes