+++ packages/Angular/Generic/entity-communication/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log - @memberjunction/ng-entity-communications
 
+## 2.70.0
+
+### Patch Changes
+
+- Updated dependencies [6f74409]
+- Updated dependencies [c9d86cd]
+  - @memberjunction/global@2.70.0
+  - @memberjunction/ng-container-directives@2.70.0
+  - @memberjunction/communication-types@2.70.0
+  - @memberjunction/entity-communications-base@2.70.0
+  - @memberjunction/entity-communications-client@2.70.0
+  - @memberjunction/core@2.70.0
+  - @memberjunction/core-entities@2.70.0
+  - @memberjunction/ng-shared@2.70.0
+
 ## 2.69.1
 
 ### Patch Changes