+++ packages/Communication/entity-comm-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @memberjunction/entity-communications-client
 
+## 2.70.0
+
+### Patch Changes
+
+- Updated dependencies [6f74409]
+- Updated dependencies [c9d86cd]
+  - @memberjunction/global@2.70.0
+  - @memberjunction/entity-communications-base@2.70.0
+  - @memberjunction/graphql-dataprovider@2.70.0
+  - @memberjunction/core@2.70.0
+  - @memberjunction/core-entities@2.70.0
+
 ## 2.69.1
 
 ### Patch Changes