+++ packages/MJDataContextServer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log - @memberjunction/data-context-server
 
+## 2.70.0
+
+### Patch Changes
+
+- Updated dependencies [6f74409]
+- Updated dependencies [c9d86cd]
+  - @memberjunction/global@2.70.0
+  - @memberjunction/data-context@2.70.0
+
 ## 2.69.1
 
 ### Patch Changes